+++ mobile/android/android-components/components/concept/storage/src/main/java/mozilla/components/concept/storage/BookmarksStorage.kt
@@ -35,6 +35,14 @@ interface BookmarksStorage : Storage {
     suspend fun getBookmarksWithUrl(url: String): List<BookmarkNode>
 
     /**
+     * Produces a list of the most recently added bookmarks.
+     *
+     * @param limit The maximum number of entries to return.
+     * @return The list of bookmarks that have been recently added up to the limit number of items.
+     */
+    suspend fun getRecentBookmarks(limit: Int): List<BookmarkNode>
+
+    /**
      * Searches bookmarks with a query string.
      *
      * @param query The query string to search.
@@ -104,7 +112,15 @@ interface BookmarksStorage : Storage {
 }
 
 /**
- * Class for holding metadata about any bookmark node
+ * Represents a bookmark record.
+ *
+ * @property type The [BookmarkNodeType] of this record.
+ * @property guid The id.
+ * @property parentGuid The id of the parent node in the tree.
+ * @property position The position of this node in the tree.
+ * @property title A title of the page.
+ * @property url The url of the page.
+ * @property children The list of children of this bookmark node in the tree.
  */
 data class BookmarkNode(
     val type: BookmarkNodeType,